[GB200NVL] check incomplete session
Whenever a session from RAKP12 times out,
we send a PAM authentication with an empty password.
If a response is received from the remote side and
it reaches the RAKP34 phase, we send a
PAM authentication with the correct password.
This allows user lockout to work and aligns with the Redfish mechanism.

Enhancement:
1. Add redfish logging event when user lockup.
2. Remove checking userData size.
3. Remove unused structure.
4. Put the ++it in the for loop
5. Use maybe_unused on userName when PAM_AUTHENTICATE is not enabled

Fixes nvbug https://nvbugspro.nvidia.com/bug/5111389

Signed-off-by: leonhuang@nvidia.com